refactor(mcp): split mcp_gateway.rs into multiple modules
